feat: #2. adds maven, and successfully builds the CI skeleton with mvn install command. updates gitignore for maven
